requirements/docs.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --annotation-style=line requirements/docs.in
#
alabaster==0.7.12         # via sphinx
babel==2.9.1              # via sphinx
certifi==2021.10.8        # via -r requirements/runtime.in, requests
charset-normalizer==2.0.11  # via requests
docutils==0.17.1          # via sphinx, sphinx-rtd-theme
googleapis-common-protos==1.54.0  # via -r requirements/runtime.in
h2==4.1.0                 # via -r requirements/../setup.txt
hpack==4.0.0              # via -r requirements/../setup.txt, h2
hyperframe==6.0.1         # via -r requirements/../setup.txt, h2
idna==3.3                 # via requests
imagesize==1.3.0          # via sphinx
importlib-metadata==4.10.1  # via sphinx
jinja2==3.0.3             # via sphinx
markupsafe==2.0.1         # via jinja2
multidict==6.0.2          # via -r requirements/../setup.txt
packaging==21.3           # via sphinx
protobuf==3.19.4          # via -r requirements/runtime.in, googleapis-common-protos
pygments==2.11.2          # via sphinx
pyparsing==3.0.7          # via packaging
pytz==2021.3              # via babel
requests==2.27.1          # via sphinx
snowballstemmer==2.2.0    # via sphinx
sphinx==4.4.0             # via -r requirements/docs.in, sphinx-rtd-theme
sphinx-rtd-theme==1.0.0   # via -r requirements/docs.in
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==2.0.0  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.5  # via sphinx
typing-extensions==4.0.1  # via importlib-metadata
urllib3==1.26.8           # via requests
zipp==3.7.0               # via importlib-metadata
